backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
cors
python-dotenv
supabase
postgrest
httpx
